Moved angle functions to own module.
